Fix the 'Analyze' button in the UI
This commit introduces the method '$.postJSON' which is a wrapper around
'$.ajax' to be able to POST data to the backend using JSON with
content-type 'application/json'.

Moreover, this commit changes the data input from 'url' to urls, as well
as allow the ability for the user to specify multiple URLs to analyze.

Unfortunately, this commit doesn't resolve how to get the data back from
the backend yet.