[NCL-5611] Add support for Remote Infinispan server
This commit adds support for remote infinispan server, in addition to
the current embedded infinispan setup.

Configuration
=============

The switch between remote and embedded is controlled in Quarkus config
by setting the `infinispan-mode` key with either value:

- EMBEDDED (default)
- REMOTE

If `REMOTE` is set, the following Quarkus config keys also need to be
configured:

- quarkus.infinispan-client.server-list (comma-delimited)
- quarkus.infinispan-client.auth-username
- quarkus.infinispan-client.auth-password

They are used to configure to which Infinispan remote server(s) to talk
to. Those keys are also used by the infinispan-quarkus-client, which we
plan to use later if we deprecate the Embedded mode.

Remote
------
When using the remote infinispan-mode, these caches need to be present
in the remote Infinispan server:

- builds
- builds-pnc
- checksums-md5
- checksums-pnc-md5
- checksums-pnc-sha1
- checksums-pnc-sha256
- checksums-sha1
- checksums-sha256
- files-md5
- files-sha1
- files-sha256
- rpms-md5
- rpms-sha1
- rpms-sha256