[NCL-6344] Do not confuse base buildConfig and provided
We use the `BuildConfig` class both for the default configuration, and
for analysis.

Inside `Finder.java`, we overload the 'config' name inside some methods
(both field and parameter are named 'config'). To avoid any confusion,
this commit renames the field 'config' to 'defaultConfig'.

One potential bug it avoids is that we could accidentally call
`cleaner.clean` on the wrong 'config', causing great disaster.