[NCL-6344] Remove commented out cleanup code
It was assumed that either the build-finder or the distribution-analyzer
code would create contents in the filesystem after each run. However,
they both only write contents to the filesystem if explicitly asked (via
`outputToFile` method), and do not write anything to the filesystem by
default.

The download of the zip to analyze is done via commons-vfs, which is an
abstraction of a filesystem and is cleaned up as needed by the garbage
collector.

Therefore, this commit removes the commented out cleanup code since they
are really not needed.