[NCL-6204] Enable AnalyzeResourceWithMockedBrewTest
For some reason, Quarkus doesn't run tests if the test class doesn't end
with 'Test'. This is why the test class got renamed.

Enabling that test class flagged all kinds of issues.

- PNC requests not matching for wiremock. This is fixed by adding the
  query for builds not being null in the request files
- PNC requests also not matching because of the scenarios not matching.
  This is fixed by just removing the scenarios from the request files
- Brew requests not matching for wiremock. This is caused by the
  checksum order in the request being random. I "fixed" it by ignoring
  the checksum value in the request

- `AnalyzerResourceTestWithDummyBrew` has a dummy Brew client session
  implementation that was being injected in
  `AnalyzeResourceWithMockedBrewTest`. I had to remove the annotations
  in the dummy Brew client session to get it to work. This is a TODO
  item.

With those changes, the tests now pass. There will be a future commit to
fix the `AnalyzeResourceTestWithDummyBrew`.

The existing `AnalyzeResourceWithMockedBrewTest` is used to verify on
the basic callback functionality of the endpoint.